Refactor to enable user supplied preProc functions for nodes/edges
User can now supply a function that is mapped to each node/edge object in supplied graph to return a new node/edge object with fields as required by vis-js. Enables more easier re-use of this component.

NOTE: background color of nodes are still processed within the component based on the 'type' of the node. This can be overridden by setting the `color` field in the node object since the internal color processing happens prior to the nodePreProc hook being fired.